Split IC.Ref.Spec
this module has grown large, and it is nice to be able to use the
“agent” functionality a bit more easily, so this splits it.
