+++ src/IC/Test/Spec.hs
@@ -4,286 +4,61 @@ This module contains a test suite for the Internet Computer
 
 -}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE ImplicitParams #-}
 {-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE LambdaCase #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE OverloadedLabels #-}
-{-# LANGUAGE TypeOperators #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE NumericUnderscores #-}
-{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE TupleSections #-}
 {-# LANGUAGE NamedFieldPuns #-}
-{-# LANGUAGE QuantifiedConstraints #-}
-{-# LANGUAGE AllowAmbiguousTypes #-}
-{-# OPTIONS_GHC -Wno-orphans #-}
-module IC.Test.Spec (preFlight, TestConfig, connect, ReplWrapper(..), icTests) where
+
+module IC.Test.Spec (icTests) where
 
 import qualified Data.Text as T
 import qualified Data.Text.Encoding as T
 import qualified Data.Text.Encoding.Error as T
-import qualified Text.Hex as H
 import qualified Data.ByteString.Lazy as BS
-import qualified Data.ByteString.Builder as BS
 import qualified Data.HashMap.Lazy as HM
 import qualified Data.Map.Lazy as M
 import qualified Data.Set as S
 import qualified Data.Vector as Vec
-import Network.HTTP.Client
-import Network.HTTP.Client.TLS
-import Network.HTTP.Types
 import Numeric.Natural
 import Data.List
-import Data.Char
 import Test.Tasty
 import Test.Tasty.HUnit
-import Test.Tasty.Options
-import Control.Monad.Trans
-import Control.Concurrent
 import Control.Monad
-import Control.Exception (catch)
-import Data.Traversable
 import Data.Word
 import Data.Functor
-import GHC.TypeLits
 import System.FilePath
 import System.Directory
 import System.Environment
-import System.Random
-import System.Exit
+import Network.HTTP.Client
 import Data.Time.Clock.POSIX
-import Codec.Candid (Principal(..), prettyPrincipal)
 import qualified Data.Binary.Get as Get
+import Codec.Candid (Principal(..))
 import qualified Codec.Candid as Candid
 import Data.Row
-import qualified Data.Row.Records as R
-import qualified Data.Row.Variants as V
-import qualified Data.Row.Internal as R
 
 import IC.Types (EntityId(..))
-import IC.Version
 import IC.HTTP.GenR
-import IC.HTTP.GenR.Parse
-import IC.HTTP.CBOR (decode, encode)
 import IC.HTTP.RequestId
+import qualified IC.HTTP.CBOR as CBOR
 import IC.Management
 import IC.Crypto
 import qualified IC.Crypto.CanisterSig as CanisterSig
 import qualified IC.Crypto.DER as DER
-import qualified IC.Crypto.DER_BLS as DER_BLS
 import IC.Id.Forms hiding (Blob)
-import IC.Test.Options
 import IC.Test.Universal
 import IC.HashTree hiding (Blob, Label)
 import IC.Certificate
-import IC.Certificate.Value
-import IC.Certificate.CBOR
 import IC.Hash
+import IC.Test.Agent
 
--- * Test data, standard requests
+type Blob = BS.ByteString
 
-doesn'tExist :: Blob
-doesn'tExist = "\xDE\xAD\xBE\xEF" -- hopefully no such canister/user exists
+-- * The test suite (see below for helper functions)
 
-defaultSK :: SecretKey
-defaultSK = createSecretKeyEd25519 "fixed32byteseedfortesting"
-
-otherSK :: SecretKey
-otherSK = createSecretKeyEd25519 "anotherfixed32byteseedfortesting"
-
-webAuthnSK :: SecretKey
-webAuthnSK = createSecretKeyWebAuthn "webauthnseed"
-
-ecdsaSK :: SecretKey
-ecdsaSK = createSecretKeyECDSA "ecdsaseed"
-
-secp256k1SK :: SecretKey
-secp256k1SK = createSecretKeySecp256k1 "secp256k1seed"
-
-defaultUser :: Blob
-defaultUser = mkSelfAuthenticatingId $ toPublicKey defaultSK
-otherUser :: Blob
-otherUser = mkSelfAuthenticatingId $ toPublicKey otherSK
-webAuthnUser :: Blob
-webAuthnUser = mkSelfAuthenticatingId $ toPublicKey webAuthnSK
-ecdsaUser :: Blob
-ecdsaUser = mkSelfAuthenticatingId $ toPublicKey ecdsaSK
-secp256k1User :: Blob
-secp256k1User = mkSelfAuthenticatingId $ toPublicKey secp256k1SK
-anonymousUser :: Blob
-anonymousUser = "\x04"
-
-queryToNonExistant :: GenR
-queryToNonExistant = rec
-    [ "request_type" =: GText "query"
-    , "sender" =: GBlob anonymousUser
-    , "canister_id" =: GBlob doesn'tExist
-    , "method_name" =: GText "foo"
-    , "arg" =: GBlob "nothing to see here"
-    ]
-
-readStateEmpty :: GenR
-readStateEmpty = rec
-    [ "request_type" =: GText "read_state"
-    , "sender" =: GBlob defaultUser
-    , "paths" =: GList []
-    ]
-
-trivialWasmModule :: Blob
-trivialWasmModule = "\0asm\1\0\0\0"
-
-addIfNotThere :: Monad m => T.Text -> m GenR -> GenR -> m GenR
-addIfNotThere f _ (GRec hm)| f `HM.member` hm = return (GRec hm)
-addIfNotThere f a (GRec hm) = do
-  x <- a
-  return $ GRec $ HM.insert f x hm
-addIfNotThere _ _ _ = error "addIfNotThere: not a record"
-
-deleteField :: T.Text -> GenR -> GenR
-deleteField f (GRec hm) = GRec $ HM.delete f hm
-deleteField _ _ = error "deleteField: not a record"
-
-modNatField :: T.Text -> (Natural -> Natural) -> GenR -> GenR
-modNatField f g (GRec hm) = GRec $ HM.adjust underNat f hm
-  where underNat :: GenR -> GenR
-        underNat (GNat n) = GNat (g n)
-        underNat _ = error "modNatField: not a nat field"
-modNatField _ _ _ = error "modNatField: not a record"
-
-addNonce :: GenR -> IO GenR
-addNonce = addIfNotThere "nonce" $
-    GBlob <$> getRand8Bytes
-
--- Adds expiry 1 minute
-addExpiry :: GenR -> IO GenR
-addExpiry = addIfNotThere "ingress_expiry" $ do
-    t <- getPOSIXTime
-    return $ GNat $ round ((t + 60) * 1000_000_000)
-
-envelopeFor :: Blob -> GenR -> IO GenR
-envelopeFor u content | u == anonymousUser = return $ rec [ "content" =: content ]
-envelopeFor u content = envelope key content
-  where
-    key ::  SecretKey
-    key | u == defaultUser = defaultSK
-        | u == otherUser = otherSK
-        | u == webAuthnUser = webAuthnSK
-        | u == ecdsaUser = ecdsaSK
-        | u == secp256k1User = secp256k1SK
-        | u == anonymousUser = error "No key for the anonymous user"
-        | otherwise = error $ "Don't know key for user " ++ show u
-
-envelope :: SecretKey -> GenR -> IO GenR
-envelope sk = delegationEnv sk []
-
-delegationEnv :: SecretKey -> [(SecretKey, Maybe [Blob])] -> GenR -> IO GenR
-delegationEnv sk1 dels content = do
-    let sks = sk1 : map fst dels
-
-    t <- getPOSIXTime
-    let expiry = round ((t + 60) * 1000_000_000)
-    delegations <- for (zip sks dels) $ \(sk1, (sk2,targets)) -> do
-      let delegation = rec $
-            [ "pubkey" =: GBlob (toPublicKey sk2)
-            , "expiration" =: GNat expiry
-            ] ++
-            [ "targets" =: GList (map GBlob ts) | Just ts <- pure targets ]
-      sig <- sign "ic-request-auth-delegation" sk1 (requestId delegation)
-      return $ rec [ "delegation" =: delegation, "signature" =: GBlob sig ]
-    sig <- sign "ic-request" (last sks) (requestId content)
-    return $ rec $
-      [ "sender_pubkey" =: GBlob (toPublicKey sk1)
-      , "sender_sig" =: GBlob sig
-      , "content" =: content
-      ] ++
-      [ "sender_delegation" =: GList delegations | not (null delegations) ]
-
--- a little bit of smartness in our combinators
-
-senderOf :: GenR -> Blob
-senderOf (GRec hm) | Just (GBlob id) <- HM.lookup "sender" hm = id
-senderOf _ = anonymousUser
-
-addNonceExpiryEnv :: GenR -> IO GenR
-addNonceExpiryEnv req = do
-  addNonce req >>= addExpiry >>= envelopeFor (senderOf req)
-
-badEnvelope :: GenR -> GenR
-badEnvelope content = rec
-    [ "sender_pubkey" =: GBlob (toPublicKey defaultSK)
-    , "sender_sig" =: GBlob (BS.replicate 64 0x42)
-    , "content" =: content
-    ]
-
-noDomainSepEnv :: SecretKey -> GenR -> IO GenR
-noDomainSepEnv sk content = do
-  sig <- sign "" sk (requestId content)
-  return $ rec
-    [ "sender_pubkey" =: GBlob (toPublicKey sk)
-    , "sender_sig" =: GBlob sig
-    , "content" =: content
-    ]
-
-noExpiryEnv, pastExpiryEnv, futureExpiryEnv :: GenR -> GenR
-noExpiryEnv = deleteField "ingress_expiry"
-pastExpiryEnv = modNatField "ingress_expiry" (subtract 3600_000_000_000)
-futureExpiryEnv = modNatField "ingress_expiry" (+ 3600_000_000_000)
-
--- * Preflight checks: Get the root key, and tell user about versions
-
-data TestConfig = TestConfig
-    { tc_root_key :: Blob
-    , tc_manager :: Manager
-    , tc_endPoint :: String
-    }
-
-makeTestConfig :: String -> IO TestConfig
-makeTestConfig ep' = do
-    manager <- newTlsManagerWith $ tlsManagerSettings
-      { managerResponseTimeout = responseTimeoutMicro 60_000_000 -- 60s
-      }
-    request <- parseRequest $ ep ++ "/api/v2/status"
-    putStrLn $ "Fetching endpoint status from " ++ show ep ++ "..."
-    s <- (httpLbs request manager >>= okCBOR >>= statusResonse)
-        `catch` (\(HUnitFailure _ r) -> putStrLn r >> exitFailure)
-
-    putStrLn $ "Spec version tested:  " ++ T.unpack specVersion
-    putStrLn $ "Spec version claimed: " ++ T.unpack (status_api_version s)
-
-    return TestConfig
-        { tc_root_key = status_root_key s
-        , tc_manager = manager
-        , tc_endPoint = ep
-        }
-  where
-    -- strip trailing slash
-    ep | null ep'        = error "empty endpoint"
-       | last ep' == '/' = init ep'
-       | otherwise       = ep'
-
-preFlight :: OptionSet -> IO TestConfig
-preFlight os = do
-    let Endpoint ep = lookupOption os
-    makeTestConfig ep
-
-
-newtype ReplWrapper = R (forall a. (HasTestConfig => a) -> a)
--- |  This is for use from the Haskell REPL, see README.md
-connect :: String -> IO ReplWrapper
-connect ep = do
-    testConfig <- makeTestConfig ep
-    let ?testConfig = testConfig
-    return (R id)
-
-
--- * The actual test suite (see below for helper functions)
-
-icTests :: TestConfig -> TestTree
-icTests = withTestConfig $ testGroup "Interface Spec acceptance tests"
+icTests :: AgentConfig -> TestTree
+icTests = withAgentConfig $ testGroup "Interface Spec acceptance tests"
   [ simpleTestCase "create and install" $ \_ ->
       return ()
 
@@ -2341,72 +2116,74 @@ icTests = withTestConfig $ testGroup "Interface Spec acceptance tests"
     ]
   ]
 
-type Blob = BS.ByteString
+-- * Equality assertions
 
--- * HUnit error reporting integration
+is :: (HasCallStack, Eq a, Show a) => a -> a -> Assertion
+is exp act = act @?= exp
 
--- To use IC.HTTP.CBOR with HUnit
-instance Parse IO where parseError = assertFailure . T.unpack
+isSet :: (HasCallStack, Ord a, Show a) => [a] -> [a] -> Assertion
+isSet exp act = S.fromList exp @?= S.fromList act
 
-asRight :: HasCallStack => Either T.Text a -> IO a
-asRight (Left err) = assertFailure (T.unpack err)
-asRight (Right gr) = return gr
+-- * Simple Wasm
 
--- * Requests
+trivialWasmModule :: Blob
+trivialWasmModule = "\0asm\1\0\0\0"
 
--- | Posting a CBOR request, returning a raw bytestring
-postCBOR :: (HasCallStack, HasTestConfig) => String -> GenR -> IO (Response BS.ByteString)
-postCBOR path gr = do
-    request <- parseRequest $ endPoint ++ path
-    request <- return $ request
-      { method = "POST"
-      , requestBody = RequestBodyLBS $ BS.toLazyByteString $ encode gr
-      , requestHeaders = [(hContentType, "application/cbor")]
-      }
-    httpLbs request testManager
+-- * Some test data related to standard requests
 
--- | postCBOR with url based on effective canister id
-postCallCBOR, postQueryCBOR, postReadStateCBOR :: (HasCallStack, HasTestConfig) => Blob -> GenR -> IO (Response BS.ByteString)
-postCallCBOR cid      = postCBOR $ "/api/v2/canister/" ++ textual cid ++ "/call"
-postQueryCBOR cid     = postCBOR $ "/api/v2/canister/" ++ textual cid ++ "/query"
-postReadStateCBOR cid = postCBOR $ "/api/v2/canister/" ++ textual cid ++ "/read_state"
+queryToNonExistant :: GenR
+queryToNonExistant = rec
+    [ "request_type" =: GText "query"
+    , "sender" =: GBlob anonymousUser
+    , "canister_id" =: GBlob doesn'tExist
+    , "method_name" =: GText "foo"
+    , "arg" =: GBlob "nothing to see here"
+    ]
 
--- | Add envelope to CBOR request, add a nonce and expiry if it is not there,
--- post to "read", return decoded CBOR
-queryCBOR :: (HasCallStack, HasTestConfig) => Blob -> GenR -> IO GenR
-queryCBOR cid req = do
-  addNonceExpiryEnv req >>= postQueryCBOR cid >>= okCBOR
+readStateEmpty :: GenR
+readStateEmpty = rec
+    [ "request_type" =: GText "read_state"
+    , "sender" =: GBlob defaultUser
+    , "paths" =: GList []
+    ]
 
--- | Add envelope to CBOR, and a nonce and expiry if not there, post to
--- "submit". Returns either a HTTP Error code, or if the status is 2xx, poll
--- for the request response, and return decoded CBOR
-type HTTPErrOrReqResponse = Either (Int,String) ReqResponse
-awaitCall' :: (HasCallStack, HasTestConfig) => Blob -> GenR -> IO HTTPErrOrReqResponse
-awaitCall' cid req = do
-  req <- addNonce req
-  req <- addExpiry req
-  res <- envelopeFor (senderOf req) req >>= postCallCBOR cid
-  let code = statusCode (responseStatus res)
-  if 200 <= code && code < 300
-  then do
-     assertBool "Response body not empty" (BS.null (responseBody res))
-     Right <$> awaitStatus (senderOf req) cid (requestId req)
-  else do
-    let msg = T.unpack (T.decodeUtf8With T.lenientDecode (BS.toStrict (BS.take 200 (responseBody res))))
-    pure $ Left (code, msg)
+badEnvelope :: GenR -> GenR
+badEnvelope content = rec
+    [ "sender_pubkey" =: GBlob (toPublicKey defaultSK)
+    , "sender_sig" =: GBlob (BS.replicate 64 0x42)
+    , "content" =: content
+    ]
 
--- | Add envelope to CBOR, and a nonce and expiry if not there, post to
--- "submit", poll for the request response, and return decoded CBOR
-awaitCall :: (HasCallStack, HasTestConfig) => Blob -> GenR -> IO ReqResponse
-awaitCall cid req = awaitCall' cid req >>= is2xx
+-- * Bad requests
 
-is2xx :: HasCallStack => HTTPErrOrReqResponse -> IO ReqResponse
-is2xx = \case
-    Left (c,msg) -> assertFailure $ "Status " ++ show c ++ " is not 2xx:\n" ++ msg
-    Right res -> pure res
+noDomainSepEnv :: SecretKey -> GenR -> IO GenR
+noDomainSepEnv sk content = do
+  sig <- sign "" sk (requestId content)
+  return $ rec
+    [ "sender_pubkey" =: GBlob (toPublicKey sk)
+    , "sender_sig" =: GBlob sig
+    , "content" =: content
+    ]
 
--- | Submits twice
-awaitCallTwice :: HasTestConfig => Blob -> GenR -> IO ReqResponse
+noExpiryEnv, pastExpiryEnv, futureExpiryEnv :: GenR -> GenR
+noExpiryEnv = deleteField "ingress_expiry"
+pastExpiryEnv = modNatField "ingress_expiry" (subtract 3600_000_000_000)
+futureExpiryEnv = modNatField "ingress_expiry" (+ 3600_000_000_000)
+
+deleteField :: T.Text -> GenR -> GenR
+deleteField f (GRec hm) = GRec $ HM.delete f hm
+deleteField _ _ = error "deleteField: not a record"
+
+modNatField :: T.Text -> (Natural -> Natural) -> GenR -> GenR
+modNatField f g (GRec hm) = GRec $ HM.adjust underNat f hm
+  where underNat :: GenR -> GenR
+        underNat (GNat n) = GNat (g n)
+        underNat _ = error "modNatField: not a nat field"
+modNatField _ _ _ = error "modNatField: not a record"
+
+-- * Double request
+
+awaitCallTwice :: HasAgentConfig => Blob -> GenR -> IO ReqResponse
 awaitCallTwice cid req = do
   req <- addNonce req
   req <- addExpiry req
@@ -2417,171 +2194,13 @@ awaitCallTwice cid req = do
   assertBool "Response body not empty" (BS.null (responseBody res))
   awaitStatus (senderOf req) cid (requestId req)
 
-getStateCert' :: (HasCallStack, HasTestConfig) => Blob -> Blob -> [[Blob]] -> IO (Response Blob)
-getStateCert' sender ecid paths = do
-    req <- addExpiry $ rec
-      [ "request_type" =: GText "read_state"
-      , "sender" =: GBlob sender
-      , "paths" =: GList (map (GList . map GBlob) paths)
-      ]
-    envelopeFor (senderOf req) req >>= postReadStateCBOR ecid
 
-decodeCert' :: HasCallStack => Blob -> IO Certificate
-decodeCert' b = either (assertFailure . T.unpack) return $ decodeCert b
-
-getStateCert :: (HasCallStack, HasTestConfig) => Blob -> Blob -> [[Blob]] -> IO Certificate
-getStateCert sender ecid paths = do
-    gr <- getStateCert' sender ecid paths >>= okCBOR
-    b <- record (field blob "certificate") gr
-    cert <- decodeCert' b
-
-    case wellFormed (cert_tree cert) of
-        Left err -> assertFailure $ "Hash tree not well formed: " ++ err
-        Right () -> return ()
-
-    return cert
-
-extractCertData :: Blob -> Blob -> IO Blob
-extractCertData cid b = do
-  cert <- decodeCert' b
-  case wellFormed (cert_tree cert) of
-      Left err -> assertFailure $ "Hash tree not well formed: " ++ err
-      Right () -> return ()
-  certValue cert ["canister", cid, "certified_data"]
-
-verboseVerify :: String -> Blob -> Blob -> Blob -> Blob -> IO ()
-verboseVerify what domain_sep pk msg sig =
-    case DER_BLS.verify domain_sep pk msg sig of
-        Left err -> assertFailure $ unlines
-            [ "Signature verification failed on " ++ what
-            , T.unpack err
-            , "Domain separator:   " ++ prettyBlob domain_sep
-            , "Public key (DER):   " ++ asHex pk
-            , "Public key decoded: " ++
-               case DER.decode pk of
-                 Left err -> T.unpack err
-                 Right (suite, key) -> asHex key ++ " (" ++ show suite ++ ")"
-            , "Signature:          " ++ asHex sig
-            , "Checked message:    " ++ prettyBlob msg
-            ]
-        Right () -> return ()
-
-validateDelegation :: (HasCallStack, HasTestConfig) => Maybe Delegation -> IO Blob
-validateDelegation Nothing = return (tc_root_key testConfig)
-validateDelegation (Just del) = do
-    cert <- decodeCert' (del_certificate del)
-    case wellFormed (cert_tree cert) of
-        Left err -> assertFailure $ "Hash tree not well formed: " ++ err
-        Right () -> return ()
-    validateStateCert' "certificate delegation" cert
-
-    certValue cert ["subnet", del_subnet_id del, "public_key"]
-
-validateStateCert' :: (HasCallStack, HasTestConfig) => String -> Certificate -> IO ()
-validateStateCert' what cert = do
-    pk <- validateDelegation (cert_delegation cert)
-    verboseVerify what "ic-state-root" pk (reconstruct (cert_tree cert)) (cert_sig cert)
-
-validateStateCert :: (HasCallStack, HasTestConfig) => Certificate -> IO ()
-validateStateCert = validateStateCert' "certificate"
-
-data ReqResponse = Reply Blob | Reject Natural T.Text
-  deriving (Eq, Show)
-data ReqStatus = Processing | Pending | Responded ReqResponse | UnknownStatus
-  deriving (Eq, Show)
-
-prettyPath :: [Blob] -> String
-prettyPath = concatMap (("/" ++) . shorten 15 . prettyBlob)
-
-prettyBlob :: Blob -> String
-prettyBlob x =
-  let s = map (chr . fromIntegral) (BS.unpack x) in
-  if all isPrint s then s else asHex x
-
-certValue :: HasCallStack => CertVal a => Certificate -> [Blob] -> IO a
-certValue cert path = case lookupPath (cert_tree cert) path of
-    Found b -> case fromCertVal b of
-      Just x -> return x
-      Nothing -> assertFailure $ "Cannot parse " ++ prettyPath path ++ " from " ++ show b
-    x -> assertFailure $ "Expected to find " ++ prettyPath path ++ ", but got " ++ show x
-
-certValueAbsent :: HasCallStack => Certificate -> [Blob] -> IO ()
-certValueAbsent cert path = case lookupPath (cert_tree cert) path of
-    Absent -> return ()
-    x -> assertFailure $ "Path " ++ prettyPath path ++ " should be absent, but got " ++ show x
-
-getRequestStatus :: (HasCallStack, HasTestConfig) => Blob -> Blob -> Blob -> IO ReqStatus
-getRequestStatus sender cid rid = do
-    cert <- getStateCert sender cid [["request_status", rid]]
-
-    case lookupPath (cert_tree cert) ["request_status", rid, "status"] of
-      Absent -> return UnknownStatus
-      Found "processing" -> return Processing
-      Found "received" -> return Pending
-      Found "replied" -> do
-        b <- certValue cert ["request_status", rid, "reply"]
-        certValueAbsent cert ["request_status", rid, "reject_code"]
-        certValueAbsent cert ["request_status", rid, "reject_message"]
-        return $ Responded (Reply b)
-      Found "rejected" -> do
-        certValueAbsent cert ["request_status", rid, "reply"]
-        code <- certValue cert ["request_status", rid, "reject_code"]
-        msg <- certValue cert ["request_status", rid, "reject_message"]
-        return $ Responded (Reject code msg)
-      Found s -> assertFailure $ "Unexpected status " ++ show s
-      -- This case should not happen with a compliant IC, but let
-      -- us be liberal here, and strict in a dedicated test
-      Unknown -> return UnknownStatus
-      x -> assertFailure $ "Unexpected request status, got " ++ show x
-
-awaitStatus :: HasTestConfig => Blob -> Blob -> Blob -> IO ReqResponse
-awaitStatus sender cid rid = loop $ pollDelay >> getRequestStatus sender cid rid >>= \case
-    Responded x -> return $ Just x
-    _ -> return Nothing
-  where
-    loop :: HasCallStack => IO (Maybe a) -> IO a
-    loop act = go (0::Int)
-      where
-        go 10000 = assertFailure "Polling timed out"
-        go n = act >>= \case
-          Just r -> return r
-          Nothing -> go (n+1)
-
-pollDelay :: IO ()
-pollDelay = threadDelay $ 10 * 1000 -- 10 milliseonds
-
--- How long to wait before checking if a request that should _not_ show up on
--- the system indeed did not show up
-ingressDelay :: IO ()
-ingressDelay = threadDelay $ 2 * 1000 * 1000 -- 2 seconds
-
-
--- * HTTP Response predicates
-
-codePred :: HasCallStack => String -> (Int -> Bool) -> Response Blob -> IO ()
-codePred expt pred response = assertBool
-    ("Status " ++ show c ++ " is not " ++ expt ++ "\n" ++ msg)
-    (pred c)
-  where
-    c = statusCode (responseStatus response)
-    msg = T.unpack (T.decodeUtf8With T.lenientDecode (BS.toStrict (BS.take 200 (responseBody response))))
-
-code2xx, code202, code4xx, code202_or_4xx  :: HasCallStack => Response BS.ByteString -> IO ()
-code2xx = codePred "2xx" $ \c -> 200 <= c && c < 300
-code202 = codePred "202" $ \c -> c == 202
-code4xx = codePred "4xx" $ \c -> 400 <= c && c < 500
-code202_or_4xx = codePred "202 or 4xx" $ \c -> c == 202 || 400 <= c && c < 500
 
 -- * CBOR decoding
 
-okCBOR :: HasCallStack => Response BS.ByteString -> IO GenR
-okCBOR response = do
-  code2xx response
-  asRight $ decode $ responseBody response
-
 asCBORBlobList :: Blob -> IO [Blob]
-asCBORBlobList blob = do 
-    decoded <- asRight $ decode $ blob
+asCBORBlobList blob = do
+    decoded <- asRight $ CBOR.decode blob
     case decoded of
         GList list -> mapM cborToBlob list
         _ -> assertFailure $ "Failed to decode as CBOR encoded list of blobs: " <> show decoded
@@ -2590,129 +2209,12 @@ cborToBlob :: GenR -> IO Blob
 cborToBlob (GBlob blob) = return blob
 cborToBlob r = assertFailure $ "Expected blob, got " <> show r
 
--- * Response predicates and parsers
+-- Interaction with aaaaa-aa via the universal canister
 
-queryResponse :: GenR -> IO ReqResponse
-queryResponse = record $ do
-    s <- field text "status"
-    case s of
-      "replied" ->
-        Reply <$> field (record (field blob "arg")) "reply"
-      "rejected" -> do
-        code <- field nat "reject_code"
-        msg <- field text "reject_message"
-        return $ Reject code msg
-      _ -> lift $ assertFailure $ "Unexpected status " ++ show s
-
-isReject :: HasCallStack => [Natural] -> ReqResponse -> IO ()
-isReject _ (Reply r) =
-  assertFailure $ "Expected reject, got reply:" ++ prettyBlob r
-isReject codes (Reject n msg) = do
-  assertBool
-    ("Reject code " ++ show n ++ " not in " ++ show codes ++ "\n" ++ T.unpack msg)
-    (n `elem` codes)
-
-isErrOrReject :: HasCallStack => [Natural] -> HTTPErrOrReqResponse -> IO ()
-isErrOrReject _codes (Left (c, msg))
-    | 400 <= c && c < 600 = return ()
-    | otherwise = assertFailure $
-        "Status " ++ show c ++ " is not 4xx or 5xx:\n" ++ msg
-isErrOrReject [] (Right _) = assertFailure "Got HTTP response, expected HTTP error"
-isErrOrReject codes (Right res) = isReject codes res
-
-
-isReply :: HasCallStack => ReqResponse -> IO Blob
-isReply (Reply b) = return b
-isReply (Reject n msg) =
-  assertFailure $ "Unexpected reject (code " ++ show n ++ "): " ++ T.unpack msg
-
--- Predicates to handle the responses from relayReply and relayReject
-isRelay :: HasCallStack => Blob -> IO ReqResponse
-isRelay = runGet $ Get.getWord32le >>= \case
-    0 -> Reply <$> Get.getRemainingLazyByteString
-    0x4c444944 -> fail "Encountered Candid when expectin relayed data. Did you forget to use isRelay?"
-    c -> do
-      msg <- Get.getRemainingLazyByteString
-      return $ Reject (fromIntegral c) (T.decodeUtf8With T.lenientDecode (BS.toStrict msg))
-
--- Convenience decoders
-asWord32 :: HasCallStack => Blob -> IO Word32
-asWord32 = runGet Get.getWord32le
-
-asWord64 :: HasCallStack => Blob -> IO Word64
-asWord64 = runGet Get.getWord64le
-
-as2Word64 :: HasCallStack => Blob -> IO (Word64, Word64)
-as2Word64 = runGet $ (,) <$> Get.getWord64le <*> Get.getWord64le
-
-bothSame :: (Eq a, Show a) => (a, a) -> Assertion
-bothSame (x,y) = x @?= y
-
-runGet :: HasCallStack => Get.Get a -> Blob -> IO a
-runGet a b = case  Get.runGetOrFail (a <* done) b of
-    Left (_,_, err) ->
-        fail $ "Could not parse " ++ show b ++ ": " ++ err
-    Right (_,_, x) -> return x
-  where
-    done = do
-        nothing_left <- Get.isEmpty
-        unless nothing_left (fail "left-over bytes")
-
-is :: (HasCallStack, Eq a, Show a) => a -> a -> Assertion
-is exp act = act @?= exp
-
-isSet :: (HasCallStack, Ord a, Show a) => [a] -> [a] -> Assertion
-isSet exp act = S.fromList exp @?= S.fromList act
-
-data StatusResponse = StatusResponse
-    { status_api_version :: T.Text
-    , status_root_key :: Blob
-    }
-
-statusResonse :: HasCallStack => GenR -> IO StatusResponse
-statusResonse = record $ do
-    v <- field text "ic_api_version"
-    _ <- optionalField text "impl_source"
-    _ <- optionalField text "impl_version"
-    _ <- optionalField text "impl_revision"
-    pk <- field blob "root_key"
-    swallowAllFields -- More fields are explicitly allowed
-    return StatusResponse {status_api_version = v, status_root_key = pk }
-
--- * Interacting with aaaaa-aa (via HTTP)
-
-{-
-The code below has some repetition. That’s because we have
-
- A) multiple ways of _calling_ the Management Canister
-    (as default user, as specific user, via canister, with or without cycles),
- B) different things we want to know
-    (just the Candid-decoded reply, or the response, or even the HTTP error)
- C) and then of course different methods (which affect response decoding)
-
-So far, there is some duplication here because of that. Eventually, this should
-be refactored so that the test can declarative pick A, B and C separately.
--}
-
--- how to reach the management canister
-type IC00 = Blob -> T.Text -> Blob -> IO ReqResponse
-
-ic00as :: (HasTestConfig, HasCallStack) => Blob -> IC00
-ic00as user ecid method_name arg = awaitCall ecid $ rec
-      [ "request_type" =: GText "call"
-      , "sender" =: GBlob user
-      , "canister_id" =: GBlob ""
-      , "method_name" =: GText method_name
-      , "arg" =: GBlob arg
-      ]
-
-ic00 :: HasTestConfig => IC00
-ic00 = ic00as defaultUser
-
-ic00via :: HasTestConfig => Blob -> IC00
+ic00via :: HasAgentConfig => Blob -> IC00
 ic00via cid = ic00viaWithCycles cid 0
 
-ic00viaWithCycles :: HasTestConfig => Blob -> Word64 -> IC00
+ic00viaWithCycles :: HasAgentConfig => Blob -> Word64 -> IC00
 ic00viaWithCycles cid cycles _ecid method_name arg =
   do call' cid $
       callNew
@@ -2723,243 +2225,6 @@ ic00viaWithCycles cid cycles _ecid method_name arg =
       callPerform
    >>= isReply >>= isRelay
 
--- A variant that allows non-200 responses to submit
-ic00as' :: HasTestConfig => Blob -> Blob -> T.Text -> Blob -> IO HTTPErrOrReqResponse
-ic00as' user cid method_name arg = awaitCall' cid $ rec
-      [ "request_type" =: GText "call"
-      , "sender" =: GBlob user
-      , "canister_id" =: GBlob ""
-      , "method_name" =: GText method_name
-      , "arg" =: GBlob arg
-      ]
-
--- Now wrapping the concrete calls
--- (using Candid.toCandidService is tricky because of all stuff like passing through the effective canister id)
---
-callIC :: forall s a b.
-  (HasCallStack, HasTestConfig) =>
-  KnownSymbol s =>
-  (a -> IO b) ~ (ICManagement IO .! s) =>
-  (Candid.CandidArg a, Candid.CandidArg b) =>
-  IC00 -> Blob -> Label s -> a -> IO b
-callIC ic00 ecid l x = do
-    r <- ic00 ecid (T.pack (symbolVal l)) (Candid.encode x) >>= isReply
-    case Candid.decode r of
-        Left err -> assertFailure $ "Candid decoding error: " ++ err
-        Right y -> pure y
-
--- The following line noise is me getting out of my way
--- to be able to use `ic_create` etc. by passing a record that contains
--- a subset of settings, without Maybe
-type family UnRec r where UnRec (R.Rec r) = r
-type PartialSettings r = (R.Forall r R.Unconstrained1, R.Map Maybe r .// UnRec Settings ≈ UnRec Settings)
-fromPartialSettings :: PartialSettings r => R.Rec r -> Settings
-fromPartialSettings r =
-    R.map' Just r .//
-    R.default' @(R.IsA R.Unconstrained1 Maybe) @(UnRec Settings) d
-  where
-    d :: forall a. R.IsA R.Unconstrained1 Maybe a => a
-    d = case R.as @R.Unconstrained1 @Maybe @a of R.As -> Nothing
-
-ic_create :: (HasCallStack, HasTestConfig, PartialSettings r) => IC00 -> Rec r -> IO Blob
-ic_create ic00 ps = do
-  r <- callIC ic00 "" #create_canister $ empty
-    .+ #settings .== Just (fromPartialSettings ps)
-  return (rawPrincipal (r .! #canister_id))
-
-ic_provisional_create ::
-    (HasCallStack, HasTestConfig, PartialSettings r) =>
-    IC00 -> Maybe Natural -> Rec r -> IO Blob
-ic_provisional_create ic00 cycles ps = do
-  r <- callIC ic00 "" #provisional_create_canister_with_cycles $ empty
-    .+ #amount .== cycles
-    .+ #settings .== Just (fromPartialSettings ps)
-  return (rawPrincipal (r .! #canister_id))
-
-ic_install :: (HasCallStack, HasTestConfig) => IC00 -> InstallMode -> Blob -> Blob -> Blob -> IO ()
-ic_install ic00 mode canister_id wasm_module arg = do
-  callIC ic00 canister_id #install_code $ empty
-    .+ #mode .== mode
-    .+ #canister_id .== Principal canister_id
-    .+ #wasm_module .== wasm_module
-    .+ #arg .== arg
-
-ic_uninstall :: (HasCallStack, HasTestConfig) => IC00 -> Blob -> IO ()
-ic_uninstall ic00 canister_id = do
-  callIC ic00 canister_id #uninstall_code $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_set_controllers :: HasTestConfig => IC00 -> Blob -> [Blob] -> IO ()
-ic_set_controllers ic00 canister_id new_controllers = do
-  callIC ic00 canister_id #update_settings $ empty
-    .+ #canister_id .== Principal canister_id
-    .+ #settings .== fromPartialSettings (#controllers .== Vec.fromList (map Principal new_controllers))
-
-ic_start_canister :: HasTestConfig => IC00 -> Blob -> IO ()
-ic_start_canister ic00 canister_id = do
-  callIC ic00 canister_id #start_canister $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_stop_canister :: HasTestConfig => IC00 -> Blob -> IO ()
-ic_stop_canister ic00 canister_id = do
-  callIC ic00 canister_id #stop_canister $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_canister_status ::
-    forall a b. (a -> IO b) ~ (ICManagement IO .! "canister_status") =>
-    HasTestConfig => IC00 -> Blob -> IO b
-ic_canister_status ic00 canister_id = do
-  callIC ic00 canister_id #canister_status $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_deposit_cycles :: HasTestConfig => IC00 -> Blob -> IO ()
-ic_deposit_cycles ic00 canister_id = do
-  callIC ic00 canister_id #deposit_cycles $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_top_up :: HasTestConfig => IC00 -> Blob -> Natural -> IO ()
-ic_top_up ic00 canister_id amount = do
-  callIC ic00 canister_id #provisional_top_up_canister $ empty
-    .+ #canister_id .== Principal canister_id
-    .+ #amount .== amount
-
-ic_delete_canister :: HasTestConfig => IC00 -> Blob -> IO ()
-ic_delete_canister ic00 canister_id = do
-  callIC ic00 canister_id #delete_canister $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_raw_rand :: HasTestConfig => IC00 -> IO Blob
-ic_raw_rand ic00 =
-  callIC ic00 "" #raw_rand ()
-
--- Primed variants return the response (reply or reject)
-callIC' :: forall s a b.
-  HasTestConfig =>
-  KnownSymbol s =>
-  (a -> IO b) ~ (ICManagement IO .! s) =>
-  Candid.CandidArg a =>
-  IC00 -> Blob -> Label s -> a -> IO ReqResponse
-callIC' ic00 ecid l x = ic00 ecid (T.pack (symbolVal l)) (Candid.encode x)
-
-ic_create' ::
-    (HasCallStack, HasTestConfig, PartialSettings r) =>
-    IC00 -> Rec r -> IO ReqResponse
-ic_create' ic00 ps = do
-  callIC' ic00 "" #create_canister $ empty
-    .+ #settings .== Just (fromPartialSettings ps)
-
-ic_provisional_create' ::
-    (HasCallStack, HasTestConfig, PartialSettings r) =>
-    IC00 -> Maybe Natural -> Rec r -> IO ReqResponse
-ic_provisional_create' ic00 cycles ps = do
-  callIC' ic00 "" #provisional_create_canister_with_cycles $ empty
-    .+ #amount .== cycles
-    .+ #settings .== Just (fromPartialSettings ps)
-
-ic_install' :: HasTestConfig => IC00 -> InstallMode -> Blob -> Blob -> Blob -> IO ReqResponse
-ic_install' ic00 mode canister_id wasm_module arg =
-  callIC' ic00 canister_id #install_code $ empty
-    .+ #mode .== mode
-    .+ #canister_id .== Principal canister_id
-    .+ #wasm_module .== wasm_module
-    .+ #arg .== arg
-
-ic_update_settings' :: (HasTestConfig, PartialSettings r) => IC00 -> Blob -> Rec r -> IO ReqResponse
-ic_update_settings' ic00 canister_id r = do
-  callIC' ic00 canister_id #update_settings $ empty
-    .+ #canister_id .== Principal canister_id
-    .+ #settings .== fromPartialSettings r
-
-ic_set_controllers' :: HasTestConfig => IC00 -> Blob -> [Blob] -> IO ReqResponse
-ic_set_controllers' ic00 canister_id new_controllers = do
-  ic_update_settings' ic00 canister_id (#controllers .== Vec.fromList (map Principal new_controllers))
-
-ic_delete_canister' :: HasTestConfig => IC00 -> Blob -> IO ReqResponse
-ic_delete_canister' ic00 canister_id = do
-  callIC' ic00 canister_id #delete_canister $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_deposit_cycles' :: HasTestConfig => IC00 -> Blob -> IO ReqResponse
-ic_deposit_cycles' ic00 canister_id = do
-  callIC' ic00 canister_id #deposit_cycles $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_top_up' :: HasTestConfig => IC00 -> Blob -> Natural -> IO ReqResponse
-ic_top_up' ic00 canister_id amount = do
-  callIC' ic00 canister_id #provisional_top_up_canister $ empty
-    .+ #canister_id .== Principal canister_id
-    .+ #amount .== amount
-
--- Double primed variants are only for requests from users (so they take the user,
--- not a generic ic00 thing), and return the HTTP error code or the response
--- (reply or reject)
-
-callIC'' :: forall s a b.
-  HasTestConfig =>
-  KnownSymbol s =>
-  (a -> IO b) ~ (ICManagement IO .! s) =>
-  Candid.CandidArg a =>
-  Blob -> Blob -> Label s -> a -> IO HTTPErrOrReqResponse
-callIC'' user ecid l x = ic00as' user ecid (T.pack (symbolVal l)) (Candid.encode x)
-
-ic_install'' :: (HasCallStack, HasTestConfig) => Blob -> InstallMode -> Blob -> Blob -> Blob -> IO HTTPErrOrReqResponse
-ic_install'' user mode canister_id wasm_module arg =
-  callIC'' user canister_id #install_code $ empty
-    .+ #mode .== mode
-    .+ #canister_id .== Principal canister_id
-    .+ #wasm_module .== wasm_module
-    .+ #arg .== arg
-
-ic_uninstall'' :: HasTestConfig => Blob -> Blob -> IO HTTPErrOrReqResponse
-ic_uninstall'' user canister_id =
-  callIC'' user canister_id #uninstall_code $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_set_controllers'' :: HasTestConfig => Blob -> Blob -> [Blob] -> IO HTTPErrOrReqResponse
-ic_set_controllers'' user canister_id new_controllers = do
-  callIC'' user canister_id #update_settings $ empty
-    .+ #canister_id .== Principal canister_id
-    .+ #settings .== fromPartialSettings (#controllers .== Vec.fromList (map Principal new_controllers))
-
-ic_start_canister'' :: HasTestConfig => Blob -> Blob -> IO HTTPErrOrReqResponse
-ic_start_canister'' user canister_id = do
-  callIC'' user canister_id #start_canister $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_stop_canister'' :: HasTestConfig => Blob -> Blob -> IO HTTPErrOrReqResponse
-ic_stop_canister'' user canister_id = do
-  callIC'' user canister_id #stop_canister $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_canister_status'' :: HasTestConfig => Blob -> Blob -> IO HTTPErrOrReqResponse
-ic_canister_status'' user canister_id = do
-  callIC'' user canister_id #canister_status $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_delete_canister'' :: HasTestConfig => Blob -> Blob -> IO HTTPErrOrReqResponse
-ic_delete_canister'' user canister_id = do
-  callIC'' user canister_id #delete_canister $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_deposit_cycles'' :: HasTestConfig => Blob -> Blob -> IO HTTPErrOrReqResponse
-ic_deposit_cycles'' user canister_id = do
-  callIC'' user canister_id #deposit_cycles $ empty
-    .+ #canister_id .== Principal canister_id
-
-ic_raw_rand'' :: HasTestConfig => Blob -> IO HTTPErrOrReqResponse
-ic_raw_rand'' user = do
-  callIC'' user "" #raw_rand ()
-
-
--- A barrier
-
--- This will stop and start all mentioned canisters. This guarantees
--- that all outstanding callbacks are handled
-barrier :: HasTestConfig => [Blob] -> IO ()
-barrier cids = do
-  mapM_ (ic_stop_canister ic00) cids
-  mapM_ (ic_start_canister ic00) cids
-
 -- * Interacting with the universal canister
 
 -- Some common operations on the universal canister
@@ -2967,47 +2232,47 @@ barrier cids = do
 -- e.g. to check for error conditions.
 -- The unprimed variant expect a reply.
 
-install' :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ReqResponse
+install' :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ReqResponse
 install' cid prog = do
   universal_wasm <- getTestWasm "universal_canister"
   ic_install' ic00 (enum #install) cid universal_wasm (run prog)
 
-installAt :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ()
+installAt :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ()
 installAt cid prog = do
   universal_wasm <- getTestWasm "universal_canister"
   ic_install ic00 (enum #install) cid universal_wasm (run prog)
 
 -- Also calls create, used default 'ic00'
-install :: (HasCallStack, HasTestConfig) => Prog -> IO Blob
+install :: (HasCallStack, HasAgentConfig) => Prog -> IO Blob
 install prog = do
     cid <- create
     installAt cid prog
     return cid
 
-create :: (HasCallStack, HasTestConfig) => IO Blob
+create :: (HasCallStack, HasAgentConfig) => IO Blob
 create = ic_provisional_create ic00 Nothing empty
 
-upgrade' :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ReqResponse
+upgrade' :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ReqResponse
 upgrade' cid prog = do
   universal_wasm <- getTestWasm "universal_canister"
   ic_install' ic00 (enum #upgrade) cid universal_wasm (run prog)
 
-upgrade :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ()
+upgrade :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ()
 upgrade cid prog = do
   universal_wasm <- getTestWasm "universal_canister"
   ic_install ic00 (enum #upgrade) cid universal_wasm (run prog)
 
-reinstall' :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ReqResponse
+reinstall' :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ReqResponse
 reinstall' cid prog = do
   universal_wasm <- getTestWasm "universal_canister"
   ic_install' ic00 (enum #reinstall) cid universal_wasm (run prog)
 
-reinstall :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ()
+reinstall :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ()
 reinstall cid prog = do
   universal_wasm <- getTestWasm "universal_canister"
   ic_install ic00 (enum #reinstall) cid universal_wasm (run prog)
 
-callRequestAs :: (HasCallStack, HasTestConfig) => Blob -> Blob -> Prog -> GenR
+callRequestAs :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> Prog -> GenR
 callRequestAs user cid prog = rec
     [ "request_type" =: GText "call"
     , "sender" =: GBlob user
@@ -3016,7 +2281,7 @@ callRequestAs user cid prog = rec
     , "arg" =: GBlob (run prog)
     ]
 
-callToQueryRequestAs :: (HasCallStack, HasTestConfig) => Blob -> Blob -> Prog -> GenR
+callToQueryRequestAs :: (HasCallStack, HasAgentConfig) => Blob -> Blob -> Prog -> GenR
 callToQueryRequestAs user cid prog = rec
     [ "request_type" =: GText "call"
     , "sender" =: GBlob user
@@ -3025,10 +2290,10 @@ callToQueryRequestAs user cid prog = rec
     , "arg" =: GBlob (run prog)
     ]
 
-callRequest :: (HasCallStack, HasTestConfig) => Blob -> Prog -> GenR
+callRequest :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> GenR
 callRequest cid prog = callRequestAs defaultUser cid prog
 
-callToQuery'' :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO HTTPErrOrReqResponse
+callToQuery'' :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO HTTPErrOrReqResponse
 callToQuery'' cid prog = awaitCall' cid $ callToQueryRequestAs defaultUser cid prog
 
 -- The following variants of the call combinator differ in how much failure they allow:
@@ -3038,23 +2303,23 @@ callToQuery'' cid prog = awaitCall' cid $ callToQueryRequestAs defaultUser cid p
 --   call   requires a reply response
 --   call_  requires a reply response with an empty blob (a common case)
 
-call'' :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO HTTPErrOrReqResponse
+call'' :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO HTTPErrOrReqResponse
 call'' cid prog = awaitCall' cid (callRequest cid prog)
 
-call' :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ReqResponse
+call' :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ReqResponse
 call' cid prog = call'' cid prog >>= is2xx
 
-call :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO Blob
+call :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO Blob
 call cid prog = call' cid prog >>= isReply
 
-call_ :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ()
+call_ :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ()
 call_ cid prog = call cid prog >>= is ""
 
-callTwice' :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ReqResponse
+callTwice' :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ReqResponse
 callTwice' cid prog = awaitCallTwice cid (callRequest cid prog)
 
 
-query' :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ReqResponse
+query' :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ReqResponse
 query' cid prog =
   queryCBOR cid >=> queryResponse $ rec
     [ "request_type" =: GText "query"
@@ -3064,14 +2329,25 @@ query' cid prog =
     , "arg" =: GBlob (run prog)
     ]
 
-query :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO Blob
+query :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO Blob
 query cid prog = query' cid prog >>= isReply
 
-query_ :: (HasCallStack, HasTestConfig) => Blob -> Prog -> IO ()
+query_ :: (HasCallStack, HasAgentConfig) => Blob -> Prog -> IO ()
 query_ cid prog = query cid prog >>= is ""
 
+-- Predicates to handle the responses from relayReply and relayReject
+isRelay :: HasCallStack => Blob -> IO ReqResponse
+isRelay = runGet $ Get.getWord32le >>= \case
+    0 -> Reply <$> Get.getRemainingLazyByteString
+    0x4c444944 -> fail "Encountered Candid when expectin relayed data. Did you forget to use isRelay?"
+    c -> do
+      msg <- Get.getRemainingLazyByteString
+      return $ Reject (fromIntegral c) (T.decodeUtf8With T.lenientDecode (BS.toStrict msg))
+
+
 -- Shortcut for test cases that just need one canister
-simpleTestCase :: (HasCallStack, HasTestConfig) => String -> (Blob -> IO ()) -> TestTree
+
+simpleTestCase :: (HasCallStack, HasAgentConfig) => String -> (Blob -> IO ()) -> TestTree
 simpleTestCase name act = testCase name $ install noop >>= act
 
 -- * Programmatic test generation
@@ -3091,27 +2367,6 @@ omitFields (GRec hm) act =
   where fields = sort $ HM.keys hm
 omitFields _ _ = error "omitFields needs a GRec"
 
-
-getRand8Bytes :: IO BS.ByteString
-getRand8Bytes = BS.pack <$> replicateM 8 randomIO
-
--- * Endpoint handling
-
--- Yes, implicit arguments are frowned upon. But they are also very useful.
-
-type HasTestConfig = (?testConfig :: TestConfig)
-
-withTestConfig :: (forall. HasTestConfig => a) -> TestConfig -> a
-withTestConfig act tc = let ?testConfig = tc in act
-
-testConfig :: HasTestConfig => TestConfig
-testConfig = ?testConfig
-
-endPoint :: HasTestConfig => String
-endPoint = tc_endPoint testConfig
-
-testManager :: HasTestConfig => Manager
-testManager = tc_manager testConfig
 
 -- * Test data access
 
@@ -3135,22 +2390,3 @@ getTestWasm :: FilePath -> IO BS.ByteString
 getTestWasm base = do
   fp <- getTestFile $ base <.> "wasm"
   BS.readFile fp
-
-
--- Convenience around Data.Row.Variants used as enums
-
-enum :: (AllUniqueLabels r, KnownSymbol l, (r .! l) ~ ()) => Label l -> Var r
-enum l = V.IsJust l ()
-
--- Other utilities
-
-asHex :: Blob -> String
-asHex = T.unpack . H.encodeHex . BS.toStrict
-
-textual :: Blob -> String
-textual = T.unpack . prettyPrincipal . Principal
-
-
-shorten :: Int -> String -> String
-shorten n s = a ++ (if null b then "" else "…")
-  where (a,b) = splitAt n s
