+++ src/ic-ref-test.hs
@@ -12,6 +12,7 @@ import Test.Tasty.Runners.Html
 import Test.Tasty.Runners
 
 import IC.Test.Options
+import IC.Test.Agent (preFlight)
 import IC.Test.Spec
 import qualified IC.Crypto.BLS as BLS
 
@@ -19,8 +20,8 @@ main :: IO ()
 main = do
     BLS.init
     os <- parseOptions ingredients (testGroup "dummy" [])
-    tc <- preFlight os
-    defaultMainWithIngredients ingredients (icTests tc)
+    ac <- preFlight os
+    defaultMainWithIngredients ingredients (icTests ac)
   where
     ingredients =
       [ rerunningTests
